Add validation to destroy resource

+++ app/Http/Controllers/Api/CalendarResourceController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\CalendarResourceResource;
+use App\Models\CalendarEvent;
 use App\Models\CalendarResource;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\ValidationException;
 
 class CalendarResourceController extends Controller
 {
@@ -80,6 +82,19 @@ class CalendarResourceController extends Controller
 
     public function destroy(CalendarResource $calendarResource)
     {
+        $user = Auth::user();
+
+        if ($calendarResource->tenant_id !== $user->tenant_id) {
+            throw ValidationException::withMessages([
+                'resource' => ['You can not delete this resource'],
+            ]);
+        }
+
+        CalendarEvent::query()
+            ->where('calendar_resource_id', $calendarResource->id)
+            ->where('tenant_id', $user->tenant_id)
+            ->delete();
+
         $calendarResource->delete();
 
         return response()->noContent();
@@ -92,9 +107,16 @@ class CalendarResourceController extends Controller
         ]);
 
         $ids = data_get($input, 'ids');
+        $user = Auth::user();
+
+        CalendarEvent::query()
+            ->whereIn('calendar_resource_id', $ids)
+            ->where('tenant_id', $user->tenant_id)
+            ->delete();
 
         CalendarResource::query()
             ->whereIn('id', $ids)
+            ->where('tenant_id', $user->tenant_id)
             ->delete();
 
         return response()->noContent();
